completed exercise 4.15-4.23

+++ controllers/blogs.js
@@ -1,20 +1,40 @@
 const blogRouter = require('express').Router()
 const Blog = require('../models/blog')
+const User = require('../models/user')
+const logger = require('../utils/logger')
+const jwt = require('jsonwebtoken')
+const middleware = require('../utils/middleware')
+
 
 blogRouter.get('/', async (request, response) => {
-  const blogs = await Blog.find({})
+  const blogs = await Blog
+    .find({}).populate('author', { username: 1, name: 1 })
   return response.json(blogs)
 })
 
-blogRouter.post('/', async (request, response) => {
-  const blog = new Blog (request.body)
+blogRouter.post('/', middleware.userExtractor, async (request, response) => {
+  const body = request.body
+
+  const user = request.user
+
+  const blog = new Blog ({
+    title: body.title,
+    author: user.id,
+    url: body.url,
+    likes: body.likes
+  })
 
   const res = await blog.save()
+  user.blogs = user.blogs.concat(res._id)
+  await user.save()
+
   return response.status(201).json(res)
 })
 
 blogRouter.put('/:id', async (request, response) => {
   const body = request.body
+
+  logger.info(body)
 
   const blog = {
     title: body.title,
@@ -32,9 +52,17 @@ blogRouter.put('/:id', async (request, response) => {
   response.json(updatedBlog)
 })
 
-blogRouter.delete('/:id', async (request, response) => {
-  await Blog.findByIdAndRemove(request.params.id)
-  response.status(204).end()
+blogRouter.delete('/:id', middleware.userExtractor, async (request, response) => {
+  const blog = await Blog.findById(request.params.id)
+
+  const user = request.user
+
+  if (blog.author.toString() === user.id) {
+    await Blog.findByIdAndRemove(request.params.id)
+    response.status(204).end()
+  } else {
+    response.status(401).json({ error: 'incorrect user' })
+  }
 })
 
 module.exports = blogRouter
